requirements/base.txt
# Django
Django>=5.1,<5.2

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Django Admin UI
django-unfold>=0.40,<1.0

# Environment
python-decouple>=3.8,<4.0

# Forms and Utils
django-widget-tweaks>=1.5,<2.0
django-crispy-forms>=2.1,<3.0
crispy-tailwind>=1.0,<2.0

# PIX QR Code
qrcode[pil]>=7.4,<8.0

# Image handling
Pillow>=10.0,<11.0

# Dependency Injection
dependency-injector>=4.41,<5.0

# Charts
django-chartjs>=2.3,<3.0
